Introduced Chart page, it has a spinner for the user to select the type of data he wants displayed.
We are using MPAndroidChart, right now more specifically the BarChart

I am noticing that the format we had at the beginning with the fragment classes and their respective .xml can be used for lots of different functionality.

Methods are quite similar as well, I am hoping for a lot of development.