Decided to change naming from WeatherData to ForecastData as it seems more appropiate for what we can do now.
The user can select the date and time and the closest forecast to the aforementioned will be displayed for the user. As the free tier of OpenWeather API doesn't allow for hourly forecast. It allows however for 3 hour, 5 days ahead. So every 3 hours for 5 days.

I've also populated ForecastData with most of the things I've found useful on the documentation site: https://openweathermap.org/forecast5

Introduced city auto-completion by providing some examples, however it is totally possible to just type in whatever city you want as long as it is in the English version (Prague, yes; Praha, no)
Their JSON format is very useful to understand how the data will be formatted when sent to us. Hence why the ForecastData class is made to resemble the format. I've done this before but now it included a lot more weather data.