Improved Budget Charts to handle multiple data on the same date
Introduced Weather Data charts, which will be available when the user fetches the data, saving them to a separate SQLite table, to have indefinetely. I figured this is way better than constantly calling the API and showing that information on the graph, as this will overload the limited API usage we have on the free tier

Introduced 3 Weather charts, temperature, humidity, clouds.

On all charts, weather or budget, the user is able to select a date range and then information between those days can be visualised.

Made it look a bit easier to understand