Added a new field pop to show the possibility of precipitation, this is very useful and I wish I would've noticed it earlier.
Also revised the rain fetching as I was noticing the rain was always wrong, I looked over the JSON API doc on OpenWeather and realized the problem probably stems from the fact that they always send a "3h" before the actual rain value, which is why I probably couldn't fetch it properly.

Added a new Chart for seeing the mm of rain. Added the functionality of fetching weather data 5 days from now with a for loop.